requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.4
